Separated logic into different packages
