+++ data/tasks.txt
@@ -1,4 +1,4 @@
-T|1|Smell
 T|1|Touch
 D|0|CS2103T|29/08/2025
-E|0|Tiktok hackathon meetup|25/08/2025 12:00|17:00
+E|1|Tiktok hackathon meetup|25/08/2025 12:00|17:00
+T|1|Smell
