+++ src/main/java/johnny/JohnnyException.java
@@ -1,3 +1,4 @@
+package johnny;
 public class JohnnyException extends Exception {
     public JohnnyException(String message) {
         super(message);
