+++ src/main/java/johnny/commands/ByeCommand.java
@@ -1,3 +1,9 @@
+package johnny.commands;
+
+import johnny.storage.Storage;
+import johnny.tasklist.TaskList;
+import johnny.ui.Ui;
+
 public class ByeCommand extends Command {
     public ByeCommand() { }
