+++ src/main/java/johnny/storage/Storage.java
@@ -1,7 +1,17 @@
+package johnny.storage;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Scanner;
+
+import johnny.parser.Parser;
+import johnny.tasklist.TaskList;
+import johnny.tasks.DeadlineTask;
+import johnny.tasks.EventTask;
+import johnny.tasks.Task;
+import johnny.tasks.TodoTask;
+import johnny.ui.Ui;
+
 import java.util.ArrayList;
 import java.util.NoSuchElementException;
 import java.time.LocalDate;
